requirements.txt
harvesters==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.2.54
Pillow==9.0.0
pyserial==3.5
torch==1.10.0
